api/pyproject.toml
[project]
name = "openadopt-api"
version = "0.1.0"
description = "Open source animal adoption platform API"
requires-python = ">=3.13"
dependencies = [
    # Core Framework
    "fastapi>=0.128",
    "uvicorn[standard]>=0.40",
    "python-multipart>=0.0",
    # Database
    "sqlalchemy>=2.0.46",
    "alembic>=1.18",
    "psycopg[binary,pool]>=3.3",
    "asyncpg>=0.31",
    # Authentication
    "pyjwt>=2.10",
    "bcrypt>=5.0",
    # Validation & Settings
    "pydantic[email]>=2.12",
    "pydantic-settings>=2.12",
    "python-dotenv>=1.2",
    # File Storage
    "aiofiles>=23.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "pytest-cov>=4.1",
    "black>=24.1",
    "ruff>=0.1",
    "mypy>=1.8",
]

[project.scripts]
create-super-admin = "app.scripts.create_super_admin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ty.environment]
root = ["./app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
